requirements.txt
beautifulsoup4
requests
psycopg2-binary
python-dotenv
Flask